comercial crud

+++ src/main/java/org/iesvdm/dao/ComercialDAO.java
@@ -9,7 +9,7 @@ public interface ComercialDAO {
 	public void create(Comercial cliente);
 	
 	public List<Comercial> getAll();
-	public Optional<Comercial>  find(int id);
+	public Optional<Comercial> find(long id);
 	public List<Comercial> findWithPedidosByCliente(long clienteId);
 	
 	public void update(Comercial cliente);
